MH-005: Refactor imports to streamline PAGE_METADATA_KEY usage and remove deprecated page-keys enum
