+++ src/app/(pages)/photo-editor/layout.tsx
@@ -1,6 +1,6 @@
 import { Metadata } from 'next';
 import { generatePageMetadata } from '@/helper';
-import { PAGE_METADATA_KEY } from '@/models/page-metadata';
+import { PAGE_METADATA_KEY } from '@/models';
 
 export const metadata: Metadata = generatePageMetadata(PAGE_METADATA_KEY.PHOTO_EDITOR);
