+++ src/models/index.ts
@@ -1,0 +1,2 @@
+export * from './page-metadata';
+export * from './menus';
